+++ client/.umirc.ts
@@ -1,15 +1,13 @@
-import { defineConfig } from 'umi';
+import { defineConfig } from '@umijs/max';
 import path from 'path'
 
 export default defineConfig({
   // 使用hash模式, 解决基于文件进行浏览时的白屏问题
-  history: {
-    type: "hash"
-  },
-  base: "./",
-  publicPath: "./",
+  history: { type: 'hash' },
+  base: "/",
+  publicPath: '/',
   hash: true,
-  chainWebpack(memo, { env, webpack, createCSSRule }) {
+  chainWebpack(memo, { env, webpack }) {
     // 设置 alias
     memo.resolve.alias.set('~', path.resolve(__dirname, '..'));
     // 设置target为electron
@@ -22,9 +20,22 @@ export default defineConfig({
     baseNavigator: true,
     baseSeparator: '-',
   },
-  nodeModulesTransform: {
-    type: 'none',
-  },
+  antd: {},
+  layout: false,
+  //  {
+  //   // https://umijs.org/docs/max/layout-menu#构建时配置
+  //   title: '稳部落',
+  //   locale: false,
+  //   menuRender: false,
+  //   menuHeaderRender: () => false,
+  //   pure: true,
+  //   layout: 'top',
+  //   siderWidth: 400,
+  //   suppressSiderWhenMenuEmpty: true
+  // },
+  // nodeModulesTransform: {
+  //   type: 'none',
+  // },
   routes: [
     { path: '/', component: '@/pages/index' },
     { path: '/customer_task', component: '@/pages/customer_task/index' },
@@ -33,5 +44,5 @@ export default defineConfig({
     { path: '/login', component: '@/pages/login/index' },
     { path: '/manage', component: '@/pages/manage/index' },
   ],
-  fastRefresh: {},
+  fastRefresh: true,
 });