+++ src/config/common.ts
@@ -4,7 +4,7 @@ let packageJson = CommonUtil.getPackageJsonConfig()
 let version = parseFloat(packageJson.version)
 class Common {
   static readonly version = version
-  // 感谢github-page, 免费cdn
-  static readonly checkUpgradeUri = 'https://api.yaozeyuan.online/stablog/version'
+  // 感谢jsdeliver, 免费cdn
+  static readonly checkUpgradeUri = 'https://cdn.jsdelivr.net/gh/YaoZeyuan/stablog@master/upgrade_config/version.json'
 }
 export default Common